requirements.txt
langchain>=0.1.0
langchain-core>=0.1.10
langchain-community>=0.0.20
langchain-openai>=0.0.5
langgraph>=0.0.20
openai>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0
